Moved services to separate directory - Moved `commHelper`, `UnifiedDataLoader` into a new `/services/`  directory - Updated imports on other files
