+++ www/js/diary/services.js
@@ -3,8 +3,8 @@
 import angular from 'angular';
 import { SurveyOptions } from '../survey/survey';
 import { getConfig } from '../config/dynamicConfig';
-import { getRawEntries } from '../commHelper';
-import { getUnifiedDataForInterval } from '../unifiedDataLoader'
+import { getRawEntries } from '../services/commHelper';
+import { getUnifiedDataForInterval } from '../services/unifiedDataLoader'
 
 angular.module('emission.main.diary.services', ['emission.plugin.logger',
                                                 'emission.services'])
