+++ www/js/onboarding/SaveQrPage.tsx
@@ -11,7 +11,7 @@ import QrCode, { shareQR } from "../components/QrCode";
 import { onboardingStyles } from "./OnboardingStack";
 import { preloadDemoSurveyResponse } from "./SurveyPage";
 import { storageSet } from "../plugin/storage";
-import { registerUser } from "../commHelper";
+import { registerUser } from "../services/commHelper";
 
 const SaveQrPage = ({  }) => {
