+++ www/js/services/commHelper.ts
@@ -1,5 +1,5 @@
 import { DateTime } from "luxon";
-import { logDebug } from "./plugin/logger";
+import { logDebug } from "../plugin/logger";
 
 /**
  * @param url URL endpoint for the request
