+++ www/js/survey/multilabel/confirmHelper.ts
@@ -1,7 +1,7 @@
 // may refactor this into a React hook once it's no longer used by any Angular screens
 
 import { getAngularService } from "../../angular-react-helper";
-import { fetchUrlCached } from "../../commHelper";
+import { fetchUrlCached } from "../../services/commHelper";
 import i18next from "i18next";
 import { logDebug } from "../../plugin/logger";
